refactor: 패키지 네이밍을 lowercase로 변경
