+++ app/src/main/AndroidManifest.xml
@@ -12,13 +12,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.view.seat_pick.SeatPickerActivity"
+            android:name=".presentation.view.seatpick.SeatPickerActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.book_complete.BookCompleteActivity"
+            android:name=".presentation.view.bookcomplete.BookCompleteActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.movie_detail.MovieDetailActivity"
+            android:name=".presentation.view.moviedetail.MovieDetailActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.view.main.MainActivity"
