+++ app/src/main/java/woowacourse/movie/presentation/view/bookcomplete/BookCompleteActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.view.book_complete
+package woowacourse.movie.presentation.view.bookcomplete
 
 import android.os.Bundle
 import android.widget.Toast
