+++ app/src/main/java/woowacourse/movie/presentation/view/main/MovieListViewHolder.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.databinding.ItemMovieListBinding
 import woowacourse.movie.model.Movie
-import woowacourse.movie.presentation.view.movie_detail.MovieDetailActivity
+import woowacourse.movie.presentation.view.moviedetail.MovieDetailActivity
 
 class MovieListViewHolder(binding: ItemMovieListBinding) : RecyclerView.ViewHolder(binding.root) {
     private val context = binding.root.context
