+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.movie_detail.MovieDetailActivity">
+    tools:context=".presentation.view.moviedetail.MovieDetailActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_movie_booking"
